style(fe): split into several components & rename
