+++ apps/frontend/app/(client)/(main)/contest/[contestId]/@tabs/qna/[qnaId]/_components/QnaContentArea.tsx
@@ -1,40 +1,39 @@
 import PersonFillIcon from '@/public/icons/person-fill.svg'
 import { format } from 'date-fns'
 import Image from 'next/image'
+import type { ReactElement } from 'react'
 import { FaClock } from 'react-icons/fa6'
 import type { QnaContent } from '../page'
-import { DeleteButton } from './DeleteButton'
 
 const maxTitleLength = 35
 
 export function QnaContentArea({
   data,
   className,
-  canDelete
+  canDelete,
+  DeleteButtonComponent
 }: {
   data: QnaContent
   className: string
   canDelete: boolean
+  DeleteButtonComponent: ReactElement
 }) {
   const TitleText = `[${data.category}] ${data.title}`
 
   return (
     <div className={`flex flex-col gap-[40px] ${className}`}>
-      {/* header 부분 */}
       <div className="flex flex-col gap-[12px]">
-        {/* No. & 제목 */}
         <div className="flex flex-col gap-[14px]">
           <div className="round-full text-color-neutral-60 flex w-fit items-center rounded-xl bg-[#F5F5F5] px-[16px] py-[4px] text-sm font-medium">
             {`No. ${data.order}`}
           </div>
           <div className="flex items-center justify-between gap-[20px]">
-            {/* TODO: 글자 수 초과 로직 컨펌 */}
             <p className="h-[36px] w-full max-w-[1078px] overflow-hidden text-2xl font-semibold">
               {TitleText.length <= maxTitleLength
                 ? TitleText
                 : TitleText.slice(0, maxTitleLength)}
             </p>
-            {canDelete && <DeleteButton subject="question" />}
+            {canDelete && DeleteButtonComponent}
           </div>
         </div>
         {/* userID & time */}
@@ -56,7 +55,6 @@ export function QnaContentArea({
           </div>
         </div>
       </div>
-      {/* content 부분 */}
       <div className="whitespace-pre-line font-normal">
         {data.content?.trim()}
       </div>
